Better handling of replica-migration when primary and replica have different cluster-allow-replica-migration
When the primary and replica cluster-allow-replica-migration configuration items are
different, the replica migration does not meet expectations.

Originally: primary not allow, replica allow. In this case, the code will migrate the
replica away, and the primary will keep as an empty primary. However, in #970, we found
a timing issue. For example, if the replica receives CLUSTER SETSLOT first and then receives
the gossip change, it will not trigger the replica-migration.

This situation has triggered new thinking. What should we do when the configuration is
inconsistent? What i think is that if the primary does not allow and the replica allow,
the replica should not migrate, and it needs to retain the primary-replica relationship
with the original primary. If the primary allow and the replica does not allow, after
the primary migrates, we also need to migrate the replica regradless the "not allow",
otherwise the replica will have no primary.

Closes #970.

Signed-off-by: Binbin <binloveplay1314@qq.com>