Add new RDBFLAGS_SLOT_SYNC to allow loading the same function from slot RDBs
In slot sync, we can load multiple slot RDBs from different nodes.
slot RDBs will contain functions, which may encounter an `already
exists` error when loading since every node will have the function
data. This is a limitation in ours slot RDB design.

This commit added a new RDBFLAGS_SLOT_SYNC flag, it is a hint that
means we are targeting slot rdb. With this flag, when loading the
function, we will treat it like FUNCTION LOAD REPLACE, so we won't
have errors and the function reached last will win (which won't be
a problem at the moment since all the function is the same in cluster).