Fix replica data loss when master re-load a slot RDB
We have three nodes, A and B doing the slot sync, A is the src
node, and B is the dst node. Before we doing the CLUSTER
SLOTFAILOVER, we adds a new C, and C is the replica of B.

In this time, if the connection between A and B breaks, in B's
views, when doing freeClient, it will call onSlotSyncClientClose
to reset the link, and then B will try to do a new slot SYNC in
cron. And then B will call delkeysNotOwnedByMySelf to delete all
the keys in that slot, which will propagate to C and C will becoma
a empty DB.

When B doing the new slot SYNC, A will generate a new slot RDB,
and B will re-load the new slot RDB. But this new RDB file will
not be propagated to node C because the master-replica connection
between B and C is normal, which will result a data loss in C.

In this commit, after the master node loads the slot RDB, we need
to disconnect all slave nodes and allow the slave nodes to fully
synchronize (we don't support slot psync).